requirements.txt
absl-py==1.4.0
appdirs==1.4.4
brotlipy==0.7.0
cachetools==5.3.1
chumpy==0.70
clip @ git+https://github.com/openai/CLIP.git
contourpy==1.0.7
cryptography
cycler==0.11.0
filelock
flit_core
fonttools==4.39.4
ftfy==6.1.1
gmpy2
google-auth==2.19.1
google-auth-oauthlib==1.0.0
grpcio==1.54.2
idna
imageio==2.31.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
Jinja2
joblib
kiwisolver==1.4.4
lazy_loader==0.2
Markdown==3.4.3
MarkupSafe
matplotlib==3.7.1
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
mpmath==1.2.1
networkx
numpy
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.4.0
plotly==5.14.1
pooch
progress==1.6
protobuf==3.20.3
psutil
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytorch3d==0.7.4
PyWavelets==1.4.1
PyYAML==6.0
regex==2023.6.3
rsa==4.9
scikit-image==0.21.0
scikit-learn
scipy==1.10.1
six
sympy
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorboardX==2.6
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl
tifffile
torch==2.0.0
torch-geometric
torchaudio==2.0.0
torchvision==0.15.0
tqdm
triton==2.0.0
typing_extensions
urllib3
wcwidth==0.2.6
Werkzeug==2.3.4
yacs==0.1.8
zipp==3.15.0
